+++ semcore/typography/src/index.d.ts
@@ -2,7 +2,7 @@ import React from 'react';
 import { PropGetterFn, UnknownProperties, Intergalactic } from '@semcore/core';
 
 import { Property } from 'csstype';
-import { Box, BoxProps } from '@semcore/flex-box';
+import { Box, BoxProps, Flex } from '@semcore/flex-box';
 import { KeyboardFocusProps } from '@semcore/utils/lib/enhances/keyboardFocusEnhance';
 
 /** @deprecated */
@@ -97,8 +97,12 @@ declare const Hint: Intergalactic.Component<'abbr', HintProps> & {
   Text: typeof Box;
 };
 
+declare const Item: Intergalactic.Component<'li', ListItemProps> & {
+  Content: typeof Flex;
+};
+
 declare const List: Intergalactic.Component<'ul', ListProps> & {
-  Item: Intergalactic.Component<'li', ListItemProps>;
+  Item: typeof Item;
 };
 
 declare const Text: Intergalactic.Component<'span', TextProps>;